nexural-backtesting-2/nexural_backtesting/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexural-backtesting"
version = "1.0.0"
description = "Professional Quantitative Finance Backtesting Platform"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Nexural Team", email = "team@nexural.com" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["quantitative-finance", "backtesting", "trading", "ai", "ml", "risk-management"]
requires-python = ">=3.9"

dependencies = [
    # Core Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "polars>=0.19.0",
    
    # Machine Learning & AI
    "scikit-learn>=1.3.0",
    "xgboost>=1.7.0",
    "lightgbm>=4.0.0",
    "optuna>=3.3.0",
    "hyperopt>=0.2.7",
    "scikit-optimize>=0.9.0",
    
    # Web Framework & API
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    
    # Database & Storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.28.0",
    
    # Caching & Messaging
    "redis>=4.6.0",
    "kafka-python>=2.0.0",
    "aiokafka>=0.8.0",
    
    # Security & Authentication
    "pyotp>=2.9.0",
    "qrcode>=7.4.0",
    "cryptography>=41.0.0",
    "passlib[bcrypt]>=1.7.0",
    "python-jose>=3.3.0",
    
    # Data Processing & Analysis  
    # "vaex>=4.16.0",  # Optional - can be installed separately if needed
    "dask>=2023.8.0",
    # "ray>=2.6.0",  # Optional - can be installed separately if needed
    
    # Monitoring & Logging
    "structlog>=23.1.0",
    "prometheus-client>=0.17.0",
    "opentelemetry-api>=1.19.0",
    "psutil>=5.9.0",
    
    # Data Validation
    "pydantic>=2.3.0",
    "pydantic-settings>=2.0.0",
    
    # Configuration & Environment
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
    "pytest-xdist>=3.3.0",
    "pytest-html>=3.2.0",
    "pytest-json-report>=1.5.0",
    "factory-boy>=3.3.0",
    "faker>=19.3.0",
    "responses>=0.23.0",
    "pytest-freezegun>=0.4.0",
    
    # Code Quality
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.3.0",
    "ruff>=0.0.284",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.2.0",
    "mkdocstrings[python]>=0.22.0",
]

all = ["nexural-backtesting[dev,docs]"]

[project.scripts]
nexural = "nexural_backtesting.cli:main"
nexural-api = "nexural_backtesting.api.server:main"

[project.urls]
Homepage = "https://github.com/nexural/nexural-backtesting"
Documentation = "https://nexural-backtesting.readthedocs.io"
Repository = "https://github.com/nexural/nexural-backtesting"
Issues = "https://github.com/nexural/nexural-backtesting/issues"
Changelog = "https://github.com/nexural/nexural-backtesting/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["nexural_backtesting"]

[tool.hatch.build.targets.sdist]
include = [
    "nexural_backtesting/",
    "tests/",
    "docs/",
    "examples/",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
]

# Testing Configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=nexural_backtesting",
    "--cov-report=html:build/htmlcov",
    "--cov-report=xml:build/coverage.xml",
    "--cov-report=term-missing",
    "--junit-xml=build/junit.xml",
    "--html=build/pytest-report.html",
    "--self-contained-html",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "e2e: End-to-end tests",
    "slow: Slow tests",
    "gpu: Tests requiring GPU",
    "network: Tests requiring network access",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::FutureWarning",
]

# Coverage Configuration
[tool.coverage.run]
source = ["nexural_backtesting"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/.*",
    "*/venv/*",
    "*/.venv/*",
    "*/build/*",
    "*/dist/*",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
show_missing = true
precision = 2

# Code Formatting
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\\.pyi?$'
extend-exclude = '''
/(
  # directories
  \\.eggs
  | \\.git
  | \\.hg
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["nexural_backtesting"]
known_third_party = [
    "pandas", "numpy", "scipy", "sklearn", "fastapi", 
    "sqlalchemy", "redis", "kafka", "pytest"
]

# Type Checking
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "numpy.*", 
    "scipy.*",
    "sklearn.*",
    "xgboost.*",
    "lightgbm.*",
    "kafka.*",
    "redis.*",
    "vaex.*",
    "dask.*",
    "ray.*",
]
ignore_missing_imports = true

# Linting
[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

# Security
[tool.bandit]
exclude_dirs = ["tests", "build", "dist"]
skips = ["B101", "B601"]  # Skip assert_used and shell=True warnings for tests
